+++ javaslang/src/main/java/javaslang/collection/List.java
@@ -7,7 +7,8 @@ package javaslang.collection;
 
 import javaslang.*;
 import javaslang.collection.List.Nil;
-import javaslang.collection.ListModule.*;
+import javaslang.collection.ListModule.Combinations;
+import javaslang.collection.ListModule.SplitAt;
 import javaslang.control.Option;
 
 import java.io.*;
@@ -684,7 +685,7 @@ public interface List<T> extends Kind1<List<?>, T>, LinearSeq<T>, Stack<T> {
     }
 
     @Override
-    default List<T> drop(long n) {
+    default List<T> drop(int n) {
         List<T> list = this;
         for (long i = n; i > 0 && !list.isEmpty(); i--) {
             list = list.tail();
@@ -693,7 +694,7 @@ public interface List<T> extends Kind1<List<?>, T>, LinearSeq<T>, Stack<T> {
     }
 
     @Override
-    default List<T> dropRight(long n) {
+    default List<T> dropRight(int n) {
         if (n <= 0) {
             return this;
         }
@@ -773,7 +774,7 @@ public interface List<T> extends Kind1<List<?>, T>, LinearSeq<T>, Stack<T> {
     }
 
     @Override
-    default Iterator<List<T>> grouped(long size) {
+    default Iterator<List<T>> grouped(int size) {
         return sliding(size, size);
     }
 
@@ -1182,7 +1183,7 @@ public interface List<T> extends Kind1<List<?>, T>, LinearSeq<T>, Stack<T> {
     }
 
     @Override
-    default List<T> slice(long beginIndex, long endIndex) {
+    default List<T> slice(int beginIndex, int endIndex) {
         if (beginIndex >= endIndex || beginIndex >= length() || isEmpty()) {
             return empty();
         } else {
@@ -1201,12 +1202,12 @@ public interface List<T> extends Kind1<List<?>, T>, LinearSeq<T>, Stack<T> {
     }
 
     @Override
-    default Iterator<List<T>> sliding(long size) {
+    default Iterator<List<T>> sliding(int size) {
         return sliding(size, 1);
     }
 
     @Override
-    default Iterator<List<T>> sliding(long size, long step) {
+    default Iterator<List<T>> sliding(int size, int step) {
         return iterator().sliding(size, step).map(List::ofAll);
     }
 
@@ -1242,7 +1243,7 @@ public interface List<T> extends Kind1<List<?>, T>, LinearSeq<T>, Stack<T> {
     }
 
     @Override
-    default Tuple2<List<T>, List<T>> splitAt(long n) {
+    default Tuple2<List<T>, List<T>> splitAt(int n) {
         if (isEmpty()) {
             return Tuple.of(empty(), empty());
         } else {
@@ -1336,7 +1337,7 @@ public interface List<T> extends Kind1<List<?>, T>, LinearSeq<T>, Stack<T> {
     }
 
     @Override
-    default List<T> take(long n) {
+    default List<T> take(int n) {
         if (n >= length()) {
             return this;
         }
@@ -1352,7 +1353,7 @@ public interface List<T> extends Kind1<List<?>, T>, LinearSeq<T>, Stack<T> {
     }
 
     @Override
-    default List<T> takeRight(long n) {
+    default List<T> takeRight(int n) {
         if (n >= length()) {
             return this;
         }
@@ -1472,12 +1473,12 @@ public interface List<T> extends Kind1<List<?>, T>, LinearSeq<T>, Stack<T> {
     }
 
     @Override
-    default List<Tuple2<T, Long>> zipWithIndex() {
+    default List<Tuple2<T, Integer>> zipWithIndex() {
         return zipWithIndex(Tuple::of);
     }
 
     @Override
-    default <U> List<U> zipWithIndex(BiFunction<? super T, ? super Long, ? extends U> mapper) {
+    default <U> List<U> zipWithIndex(BiFunction<? super T, ? super Integer, ? extends U> mapper) {
         Objects.requireNonNull(mapper, "mapper is null");
         return ofAll(iterator().zipWithIndex(mapper));
     }
