+++ javaslang/src/main/java/javaslang/collection/Set.java
@@ -142,10 +142,10 @@ public interface Set<T> extends Traversable<T>, Function1<T, Boolean> {
     <U> Set<T> distinctBy(Function<? super T, ? extends U> keyExtractor);
 
     @Override
-    Set<T> drop(long n);
+    Set<T> drop(int n);
 
     @Override
-    Set<T> dropRight(long n);
+    Set<T> dropRight(int n);
 
     @Override
     Set<T> dropUntil(Predicate<? super T> predicate);
@@ -163,7 +163,7 @@ public interface Set<T> extends Traversable<T>, Function1<T, Boolean> {
     <C> Map<C, ? extends Set<T>> groupBy(Function<? super T, ? extends C> classifier);
 
     @Override
-    Iterator<? extends Set<T>> grouped(long size);
+    Iterator<? extends Set<T>> grouped(int size);
 
     @Override
     Set<T> init();
@@ -205,10 +205,10 @@ public interface Set<T> extends Traversable<T>, Function1<T, Boolean> {
     <U> Set<U> scanRight(U zero, BiFunction<? super T, ? super U, ? extends U> operation);
 
     @Override
-    Iterator<? extends Set<T>> sliding(long size);
+    Iterator<? extends Set<T>> sliding(int size);
 
     @Override
-    Iterator<? extends Set<T>> sliding(long size, long step);
+    Iterator<? extends Set<T>> sliding(int size, int step);
 
     @Override
     Tuple2<? extends Set<T>, ? extends Set<T>> span(Predicate<? super T> predicate);
@@ -225,10 +225,10 @@ public interface Set<T> extends Traversable<T>, Function1<T, Boolean> {
     Option<? extends Set<T>> tailOption();
 
     @Override
-    Set<T> take(long n);
+    Set<T> take(int n);
 
     @Override
-    Set<T> takeRight(long n);
+    Set<T> takeRight(int n);
 
     @Override
     Set<T> takeUntil(Predicate<? super T> predicate);
@@ -252,8 +252,8 @@ public interface Set<T> extends Traversable<T>, Function1<T, Boolean> {
     <U> Set<Tuple2<T, U>> zipAll(Iterable<? extends U> that, T thisElem, U thatElem);
 
     @Override
-    Set<Tuple2<T, Long>> zipWithIndex();
+    Set<Tuple2<T, Integer>> zipWithIndex();
 
     @Override
-    <U> Set<U> zipWithIndex(BiFunction<? super T, ? super Long, ? extends U> mapper);
+    <U> Set<U> zipWithIndex(BiFunction<? super T, ? super Integer, ? extends U> mapper);
 }
